Implementa backend real de Ventas y Tesoreria
Agrega el modelo Sale/SaleItem (carrito con channel POS/MERCH) y el ledger
unificado TreasuryEntry, con sus modulos backend (sales, treasury) y
endpoints protegidos a Encargado/Dev. StockService ahora acepta un tx
opcional para que la venta descuente stock atomicamente dentro de la misma
transaccion que la crea. TreasuryService se muda a su propio modulo e
implementa de verdad los hooks que Cortes y Ventas ya invocaban
(recordIncomeFromCut/recordIncomeFromSale), con calculo de punto de
equilibrio (margen de contribucion aproximado + serie de cruce acumulado).
Conecta el panel admin a los endpoints reales reemplazando los mocks de
Ventas y Tesoreria.

+++ frontend/src/modules/tesoreria/tesoreria.schema.ts
@@ -4,7 +4,7 @@ export const movimientoFormSchema = z.object({
   tipo: z.enum(['costo_fijo', 'costo_variable', 'ingreso']),
   categoriaId: z.string().min(1, 'Selecciona una categoria'),
   categoriaNombre: z.string().min(1),
-  monto: z.coerce.number().positive('El monto debe ser mayor a 0'),
+  monto: z.coerce.number().int().positive('El monto debe ser mayor a 0'),
   fecha: z.string().min(1, 'La fecha es requerida'),
   descripcion: z.string().optional(),
 });
